refactor: inuiWindowをtsに移行
